+++ detox/local-cli/build-framework-cache.js
@@ -1,0 +1,10 @@
+const path = require('path');
+const cp = require('child_process');
+
+
+module.exports.command = 'build-framework-cache';
+module.exports.desc = 'Build Detox.framework to ~/Library/Detox. The framework cache is specific for each combination of Xcode and Detox versions';
+
+module.exports.handler = function () {
+  cp.execSync(path.join(__dirname, '../scripts/build_framework.ios.sh'), {stdio: 'inherit'});
+}
