--- detox/local-cli/detox-build-framework-cache.js
@@ -1,6 +1,0 @@
-#!/usr/bin/env node
-
-const path = require('path');
-const cp = require('child_process');
-
-cp.execSync(path.join(__dirname, '../scripts/build_framework.ios.sh'), {stdio: 'inherit'});
